requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
Django==4.2.16
idna==3.10
numpy==1.24.4
opencv-python==4.10.0.84
pillow==10.4.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
qrcode==7.4.2
razorpay==1.4.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
